+++ src/main/kotlin/com/ojajae/domain/dashbord/form/request/DashboardRequestForm.kt
@@ -4,5 +4,5 @@ import com.ojajae.domain.dashbord.entity.DashboardType
 
 class DashboardRequestForm(
     val type: DashboardType,
-    val storeId: Long?,
+    val storeId: Int?,
 )
